pyproject.toml
[project]
name = "FB2"
version = "0.3.0b1"
requires-python = ">=3.10"
authors = [{ name = "Ae-Mc", email = "ae_mc@mail.ru" }]
description = "A small package for working with FB2 files"
dependencies = ["iso639-lang>=2.6.3", "pydantic>=2.12.3"]
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/Ae-Mc/FB2"
Repository = "https://github.com/Ae-Mc/FB2"
Issues = "https://github.com/Ae-Mc/FB2/issues"

[dependency-groups]
dev = ["ruff>=0.14.1"]
